tests/wpt/tests/css/css-rhythm/parsing/block-step-insert-computed.html

<!DOCTYPE html>
<html>
<head>
<title>CSS Rhythm: block-step-insert computed values</title>
<link rel="author" title="Sammy Gill" href="sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-rhythm/#block-step-insert">
<meta name="assert" content="Checking computed values for block-step-insert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
    #target {
      font-size: 40px;
    }
</style>
</head>
<body>
<div id="target"></div>
<script>
    test_computed_value("block-step-insert", "margin-box");
    test_computed_value("block-step-insert", "padding-box");
    test_computed_value("block-step-insert", "content-box");
</script>
</body>
</html>
